pyproject.toml
[project]
name = "fastapi-mcp-app"
version = "1.0.0"
description = "FastAPI application with FastMCP integration"
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "fastmcp>=2.5.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "structlog>=25.3.0",
    "arize-phoenix>=10.0.4",
    "arize-phoenix-otel>=0.9.2",
    "openinference-instrumentation-mcp>=0.1.0",
    "openinference-instrumentation-openai_agents>=0.1.0"
]

[project.optional-dependencies]
dev = [
    "black>=23.10.0",
    "isort>=5.12.0",
    "flake8>=7.2.0",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.28.1",
    "mypy>=1.6.1",
    "types-python-jose>=3.3.4.3",
    "types-passlib>=1.7.5.20230806"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py39"
